build(swc): fix exports - node16 (from CJS) ESM (dynamic import only)
Re-configured the swc build process to export common JS for
Typescript+NodeJS users.

To get more details about the export issues you can visit this link:
https://arethetypeswrong.github.io/?p=http-errors-enhanced

The updated version of the package gets an all green in terms of the
checks presented on the link.

With that said, I have not done extensive testing to verify that
everything works as before (so the tool linked to above is the
only evidence so far that this is all good and not a regression)

I did publish a version on npm under the package name:
`http-errors-enhanced-cjs` which can be seen here
https://www.npmjs.com/package/http-errors-enhanced-cjs

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>